프로젝트 포맷팅하기

+++ back/tripsok_back/src/main/java/com/tripsok_back/repository/UserRepository.java
@@ -1,9 +1,10 @@
 package com.tripsok_back.repository;
 
-import com.tripsok_back.model.user.TripSokUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.tripsok_back.model.user.TripSokUser;
+
 @Repository
-public interface UserRepository extends JpaRepository<TripSokUser, Long> {
+public interface UserRepository extends JpaRepository<TripSokUser, Integer> {
 }
